Added library/build.gradle script.
The file pulls the compile target from the pom.xml in its parent
directory, thus ensuring that that value only needs to be changed
in one location.

Currently, this build file doesn't deal with tests. In the future,
somebody may wish to add this functionality, but at present, the
most common use-case for this build file will be for an app who
wishes to use UIL as a dependency, but build it from source.